refactor: list component 파일 명칭 정리
